Refactor 0 season parsing

+++ app/src/main/java/com/example/familymafiaapp/entities/Rating.kt
@@ -3,7 +3,7 @@ package com.example.familymafiaapp.entities
 data class Rating(
     val player: String,
     val ratingCoefficient: Int,
-    val winPoints: Float,
+    val winPoints: Int,
     val gamesPlayed: Int,
     val firstKilled: Int,
     val mvp: Float,
